requirements.txt
streamlit>=1.37.0
xai-sdk>=0.1.0
python-dotenv>=1.0.1
pytest>=8.0.0
